backend/requirements.txt
asgiref==3.6.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==3.11.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.5.0
djoser==2.1.0
drf-extra-fields==3.7.0
environs==9.5.0
factory-boy==3.2.1
Faker==11.3.0
gunicorn==20.1.0
idna==3.10
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==21.3
Pillow==11.1.0
pluggy==1.0.0
psycopg2-binary==2.9.7
py==1.11.0
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
reportlab==4.3.1
social-auth-app-django==4.0.0
social-auth-core==4.5.6
sqlparse==0.5.3
urllib3==1.26.15
tzdata==2025.1